Initial agent-side implementation of monitor::service
The monitor::service defined resource type is a single-resource generic
nagios monitoring type. It's more concise than the previous recipe of
@@nagios_service + nrpe::remote_command + sudo::command + a file
resource for the check script. Bundling these things into a single type
is intended to reduce the risk of cut and paste errors, as well as
adding a few features:

* Only need to specify things like service name, script/sudo commandline
  and so on once.
* Single attribute to specify a user to run the check using sudo.
* Single attribute to switch between server-side and nrpe check.
* Easy to assign services and hosts to service/hostgroups.
* Easy to assign logos and layout for the nagios map view.
* Generic, so it's feasible to replace Nagios with a different
  monitoring system without too much rewriting.